+++ Microsoft.Azure.Cosmos/FaultInjection/changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Features Added
 - Distributed Transactions: Adds fault injection support for distributed transactions
 - [#6004](https://github.com/Azure/azure-cosmos-dotnet-v3/pull/6004) RetryWith: Adds `RetryWith` (HTTP 449) server error injection support for Gateway mode
+- [#5969](https://github.com/Azure/azure-cosmos-dotnet-v3/pull/5969) Distributed Transactions: Adds fault injection support for distributed transactions
 
 #### Breaking Changes